[NFC] Refactor shell environment functionality from TestRunner.py into a new module
